ekg_app/requirements.txt
streamlit
streamlit-drawable-canvas
openai
Pillow
google-generativeai
